LayoutTests/http/tests/media/fairplay/fps-mse-unmuxed-key-rotation-expected.txt


PROMISE: requestMediaKeySystemAccess resolved
PROMISE: createMediaKeys resolved
FETCH: resources/cert.der OK
PROMISE: keys.setServerCertificate resolved
PROMISE: setMediaKeys() resolved
Created mediaSource
EVENT(sourceopen)
-
Appending Encrypted Video Buffer 1
Created sourceBuffer
FETCH: content/elementary-stream-video-header-keyid-2.m4v OK
EVENT(encrypted)
EVENT(message)
PROMISE: licenseResponse resolved
PROMISE: session.update() resolved
EVENT(updateend)
-
Appending Encrypted Video Payload
FETCH: content/elementary-stream-video-payload.m4v OK
EVENT(updateend)
-
Appending Encrypted Video Buffer 2
FETCH: content/elementary-stream-video-header-keyid-4.m4v OK
EVENT(encrypted)
EVENT(updateend)
EVENT(message)
PROMISE: licenseResponse resolved
PROMISE: session.update() resolved
Setting timestampOffset = buffered.end()
-
Appending Encrypted Video Payload
FETCH: content/elementary-stream-video-payload.m4v OK
EVENT(updateend)
END OF TEST
